requirements.txt
requests>=2.32.4
pathlib>=1.0.1
python-dotenv>=1.1.1
PyYAML>=6.0.1